+++ morpha/morpha/parsers/__init__.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from .irecord_parser import IRecordParser
+from .team_preview_record_parser import TeamPreviewRecordParser
 
-__all__ = ['IRecordParser']
+__all__ = ['IRecordParser',
+           'TeamPreviewRecordParser']